auth_service/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary
python-decouple==3.8
requests==2.31.0
PyJWT==2.8.0